fix: rename alerts
